ref(examples): Rename refund agent fixture
Rename the private demo runtime from foobar to refund-agent so the folder and workspace package describe their purpose. Update demo imports, docs, and path aliases to use @demo/refund-agent.
