+++ apps/demo-pi/evals/refund.eval.ts
@@ -1,5 +1,5 @@
 import { expect } from "vitest";
-import { createRefundAgent, type RefundCase } from "@demo/foobar";
+import { createRefundAgent, type RefundCase } from "@demo/refund-agent";
 import { piAiHarness } from "@vitest-evals/harness-pi-ai";
 import {
   describeEval,
