+++ examples/refund-agent/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@demo/foobar",
+  "name": "@demo/refund-agent",
   "private": true,
   "version": "0.1.0",
   "types": "./dist/index.d.ts",
