+++ packages/vitest-evals/README.md
@@ -35,7 +35,7 @@ npm install -D @vitest-evals/harness-ai-sdk
 
 ```ts
 import { expect } from "vitest";
-import { createRefundAgent } from "@demo/foobar";
+import { createRefundAgent } from "@demo/refund-agent";
 import { piAiHarness } from "@vitest-evals/harness-pi-ai";
 import {
   describeEval,
